datasets/mimic_ecg_loader.py
#!/usr/bin/env python3
"""
MIMIC-IV ECG Dataset Loader
===========================

Loader for MIMIC-IV ECG subset - critical care ECG data from the
MIMIC-IV (Medical Information Mart for Intensive Care IV) database.

Dataset Information:
- ECG recordings from critical care patients
- Associated with rich clinical metadata
- Various recording durations and sampling rates
- Integrated with MIMIC-IV clinical database
- High clinical relevance for critical care scenarios

Features:
- Integration with MIMIC-IV clinical data
- Patient timeline reconstruction
- Critical care ECG patterns
- Multi-modal data fusion (ECG + clinical notes + labs)
- Temporal analysis capabilities

Clinical Applications:
- Critical care monitoring
- Arrhythmia detection in ICU settings
- Clinical deterioration prediction
- Multi-modal clinical decision support
"""

import numpy as np
import pandas as pd
from typing import Dict, List, Tuple, Optional, Any
from dataclasses import dataclass
from pathlib import Path
import warnings

@dataclass
class MIMICConfig:
    """Configuration for MIMIC-IV ECG loader."""
    data_path: str = "/data/mimic-iv-ecg/"
    sampling_rate: int = 500
    signal_length: Optional[int] = None
    normalize_signals: bool = True
    load_clinical_data: bool = True

@dataclass  
class MIMICRecord:
    """MIMIC-IV ECG record with clinical metadata."""
    record_id: str
    subject_id: int
    hadm_id: Optional[int] = None
    ecg_data: Optional[np.ndarray] = None
    clinical_notes: Optional[str] = None
    
class MIMICECGLoader:
    """MIMIC-IV ECG dataset loader (placeholder implementation)."""
    
    def __init__(self, config: MIMICConfig = None, **kwargs):
        if config is None:
            config = MIMICConfig(**kwargs)
        self.config = config
    
    def load_patient_data(self, subject_id: int) -> Optional[MIMICRecord]:
        """Load patient ECG data (mock implementation)."""
        return MIMICRecord(
            record_id=f"mimic_{subject_id}",
            subject_id=subject_id,
            ecg_data=np.random.randn(12, 5000) if self.config.signal_length else None
        )